Split `render` from `renderToCanvas` in react
So we can use `composeStory` without depending on `renerToCanvas` and react dom etc
